moved Alias and Constant out of expressions.base into 'named' module
