+++ nineml/abstraction_layer/expressions/__init__.py
@@ -1,1 +1,2 @@
-from .base import Alias, ODE, Expression, ExpressionWithSimpleLHS, Constant
+from .base import ODE, Expression, ExpressionWithSimpleLHS
+from .named import Alias, Constant
